requirements.txt
Flask
Flask-Cors
flask-migrate
Flask-SQLAlchemy
gunicorn
google-auth
google-auth-httplib2
google-auth-oauthlib
google-api-python-client
googlemaps
pymysql
psycopg2-binary
python-dotenv
gspread
oauth2client
google-cloud-firestore
